backend/requirements.txt
fastapi
uvicorn
websockets
openai
python-dotenv
opencv-python
mediapipe
numpy